test: renamed and added some new tests covering user return to prop passed to component
